Consolidate RecentlyScannedStickersView.ViewModel into extension, restore 2 week expiration
